Extracted pin extensions to their own files
