+++ CoreBarcelona/Extensions/Glue/IMChat+Representable.swift
@@ -9,47 +9,6 @@
 import Foundation
 import IMCore
 
-@available(iOS 14, macOS 10.16, watchOS 7, *)
-extension IMPinnedConversationsController {
-    func unpin(chat: IMChat) {
-        if !pinnedConversationsContains(chat) {
-            return
-        }
-        
-        pinnedChats = pinnedChats.filter {
-            $0.pinningIdentifier != chat.pinningIdentifier
-        }
-    }
-    
-    func pin(chat: IMChat) {
-        if pinnedConversationsContains(chat) {
-            return
-        }
-        
-        var chats = pinnedChats
-        chats.append(chat)
-        
-        pinnedChats = chats
-    }
-    
-    private var pinnedChats: [IMChat] {
-        get {
-            self.pinnedConversationIdentifiers().compactMap {
-                IMChatRegistry.shared.existingChat(withPinningIdentifier: $0)
-            }
-        }
-        set {
-            #if os(macOS)
-            self.setPinnedChats(newValue, withUpdateReason: "we in this bitch")
-            #else
-            self.setPinnedConversationIdentifiers(NSOrderedSet(array: newValue.map {
-                $0.pinningIdentifier
-            }))
-            #endif
-        }
-    }
-}
-
 public extension IMChat {
     var representation: Chat {
         Chat(self)
@@ -70,24 +29,6 @@ public extension IMChat {
         }
         set {
             setValue(newValue == true ? 1 : 0, forChatProperty: "ignoreAlertsFlag")
-        }
-    }
-    
-    var pinned: Bool {
-        get {
-            if #available(iOS 14, macOS 10.16, watchOS 7, *) {
-                return isPinned
-            } else {
-                return false
-            }
-        }
-        set {
-            if newValue == pinned {
-                return
-            }
-            if #available(iOS 14, macOS 10.16, watchOS 7, *) {
-                newValue ? IMPinnedConversationsController.sharedInstance().pin(chat: self) : IMPinnedConversationsController.sharedInstance().unpin(chat: self)
-            }
         }
     }
